requirements.txt
dbus-python
PyGObject
colorthief
Pillow
requests
